Split scripts/init.ts to satisfy the 300-line file limit #487
