+++ package.json
@@ -1,5 +1,5 @@
 {
-	"version": "0.219.0",
+	"version": "0.220.0",
 	"name": "@joshuafolkken/kit",
 	"type": "module",
 	"license": "MIT",
