+++ scripts/init-actions.test.ts
@@ -1,0 +1,55 @@
+import { describe, expect, it } from 'vitest'
+import { init_actions } from './init-actions'
+
+const NPMRC = '.npmrc'
+const ESLINT = 'eslint.config.js'
+const PRETTIER = 'prettier.config.js'
+const PLAYWRIGHT = 'playwright.config.ts'
+const VITE = 'vite.config.ts'
+const TSCONFIG = 'tsconfig.json'
+const CSPELL = 'cspell.config.yaml'
+const LEFTHOOK = 'lefthook.yml'
+const VSCODE_EXTENSIONS = '.vscode/extensions.json'
+const VSCODE_SETTINGS = '.vscode/settings.json'
+
+const COMMON_TAIL_DESTINATIONS = [TSCONFIG, CSPELL, LEFTHOOK, VSCODE_EXTENSIONS, VSCODE_SETTINGS]
+
+const VANILLA_DESTINATIONS = [NPMRC, ESLINT, PRETTIER, PLAYWRIGHT, ...COMMON_TAIL_DESTINATIONS]
+
+const SVELTEKIT_DESTINATIONS = [
+	NPMRC,
+	ESLINT,
+	PRETTIER,
+	PLAYWRIGHT,
+	VITE,
+	...COMMON_TAIL_DESTINATIONS,
+]
+
+describe('init_actions.build_file_actions', () => {
+	it('returns the expected ordered destination list for vanilla projects', () => {
+		const destinations = init_actions.build_file_actions('vanilla').map((action) => action.dest)
+
+		expect(destinations).toEqual(VANILLA_DESTINATIONS)
+	})
+
+	it('inserts vite.config.ts only for sveltekit projects', () => {
+		const destinations = init_actions.build_file_actions('sveltekit').map((action) => action.dest)
+
+		expect(destinations).toEqual(SVELTEKIT_DESTINATIONS)
+	})
+
+	it('omits a merge handler only for playwright.config.ts', () => {
+		const without_merge = init_actions
+			.build_file_actions('sveltekit')
+			.filter((action) => action.merge === undefined)
+			.map((action) => action.dest)
+
+		expect(without_merge).toEqual([PLAYWRIGHT])
+	})
+
+	it('produces a non-empty create output for every action', () => {
+		for (const action of init_actions.build_file_actions('vanilla')) {
+			expect(action.create().length).toBeGreaterThan(0)
+		}
+	})
+})
